backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "markitdown[pdf]>=0.1.4",
    "pydantic-ai",
    "mistralai",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "sqlalchemy",
    "pymupdf",
    "uvicorn>=0.40.0",
    "weasyprint>=68.0",
    "logfire[fastapi,httpx,sqlalchemy]>=4.19.0",
    "rich>=14.2.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.13",
]

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests that use real external resources (PDFs, HTTP calls)",
]
testpaths = ["tests"]
filterwarnings = [
    "ignore:builtin type SwigPyPacked has no __module__ attribute:DeprecationWarning",
    "ignore:builtin type SwigPyObject has no __module__ attribute:DeprecationWarning",
    "ignore:builtin type swigvarlink has no __module__ attribute:DeprecationWarning",
]

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py313"

[tool.ruff.lint]
# Enable Pyflakes (F), pycodestyle (E, W), isort (I)
select = ["E", "F", "W", "I"]
ignore = []

[tool.ruff.lint.per-file-ignores]
"agent.py" = ["E501"]
